+++ src/collective/z3cform/select2/browser/select2widget.py
@@ -3,11 +3,16 @@ from eea.facetednavigation.widgets import ViewPageTemplateFile
 from eea.facetednavigation.widgets.select.widget import Widget as SelectWidget
 from eea.facetednavigation.widgets.select.interfaces import ISelectSchema
 
-from eea.faceted.vocabularies.utils import lowercase_text
-
+from Products.CMFPlone.utils import safe_unicode
 from zope.i18n import MessageFactory
 
 _ = MessageFactory('collective.z3cform.select2')
+
+
+def compare_lowercase(a, b):
+    a = safe_unicode(a)
+    b = safe_unicode(b)
+    return a.lower() == b.lower()
 
 
 class ISelect2Schema(ISelectSchema):
@@ -45,7 +50,7 @@ class Widget(SelectWidget):
             return False
 
         for item in default:
-            if lowercase_text(key, item) == 0:
+            if compare_lowercase(key, item):
                 return True
 
         return False